+++ src/plugins/recordTypes/person/optionLists.js
@@ -1,0 +1,165 @@
+import { defineMessages } from 'react-intl';
+
+export default {
+  personTitles: {
+    values: [
+      'Admiral',
+      'Baron',
+      'Baroness',
+      'Captain',
+      'Commander',
+      'Commodore',
+      'Count',
+      'Countess',
+      'Dame',
+      'Detective',
+      'Dr',
+      'General',
+      'Governor',
+      'Honorable',
+      'Judge',
+      'King',
+      'Lady',
+      'Lieutenant',
+      'Lord',
+      'Miss',
+      'Mr',
+      'Mrs',
+      'Ms',
+      'Prince',
+      'Princess',
+      'Professor',
+      'Queen',
+      'Reverend',
+      'Saint',
+      'Sergeant',
+      'Sir',
+    ],
+    messages: defineMessages({
+      Admiral: {
+        id: 'option.personTitles.Admiral',
+        defaultMessage: 'Admiral',
+      },
+      Baron: {
+        id: 'option.personTitles.Baron',
+        defaultMessage: 'Baron',
+      },
+      Baroness: {
+        id: 'option.personTitles.Baroness',
+        defaultMessage: 'Baroness',
+      },
+      Captain: {
+        id: 'option.personTitles.Captain',
+        defaultMessage: 'Captain',
+      },
+      Commander: {
+        id: 'option.personTitles.Commander',
+        defaultMessage: 'Commander',
+      },
+      Commodore: {
+        id: 'option.personTitles.Commodore',
+        defaultMessage: 'Commodore',
+      },
+      Count: {
+        id: 'option.personTitles.Count',
+        defaultMessage: 'Count',
+      },
+      Countess: {
+        id: 'option.personTitles.Countess',
+        defaultMessage: 'Countess',
+      },
+      Dame: {
+        id: 'option.personTitles.Dame',
+        defaultMessage: 'Dame',
+      },
+      Detective: {
+        id: 'option.personTitles.Detective',
+        defaultMessage: 'Detective',
+      },
+      Dr: {
+        id: 'option.personTitles.Dr',
+        defaultMessage: 'Dr',
+      },
+      General: {
+        id: 'option.personTitles.General',
+        defaultMessage: 'General',
+      },
+      Governor: {
+        id: 'option.personTitles.Governor',
+        defaultMessage: 'Governor',
+      },
+      Honorable: {
+        id: 'option.personTitles.Honorable',
+        defaultMessage: 'Honorable',
+      },
+      Judge: {
+        id: 'option.personTitles.Judge',
+        defaultMessage: 'Judge',
+      },
+      King: {
+        id: 'option.personTitles.King',
+        defaultMessage: 'King',
+      },
+      Lady: {
+        id: 'option.personTitles.Lady',
+        defaultMessage: 'Lady',
+      },
+      Lieutenant: {
+        id: 'option.personTitles.Lieutenant',
+        defaultMessage: 'Lieutenant',
+      },
+      Lord: {
+        id: 'option.personTitles.Lord',
+        defaultMessage: 'Lord',
+      },
+      Miss: {
+        id: 'option.personTitles.Miss',
+        defaultMessage: 'Miss',
+      },
+      Mr: {
+        id: 'option.personTitles.Mr',
+        defaultMessage: 'Mr',
+      },
+      Mrs: {
+        id: 'option.personTitles.Mrs',
+        defaultMessage: 'Mrs',
+      },
+      Ms: {
+        id: 'option.personTitles.Ms',
+        defaultMessage: 'Ms',
+      },
+      Prince: {
+        id: 'option.personTitles.Prince',
+        defaultMessage: 'Prince',
+      },
+      Princess: {
+        id: 'option.personTitles.Princess',
+        defaultMessage: 'Princess',
+      },
+      Professor: {
+        id: 'option.personTitles.Professor',
+        defaultMessage: 'Professor',
+      },
+      Queen: {
+        id: 'option.personTitles.Queen',
+        defaultMessage: 'Queen',
+      },
+      Reverend: {
+        id: 'option.personTitles.Reverend',
+        defaultMessage: 'Reverend',
+      },
+      Saint: {
+        id: 'option.personTitles.Saint',
+        defaultMessage: 'Saint',
+      },
+      Sergeant: {
+        id: 'option.personTitles.Sergeant',
+        defaultMessage: 'Sergeant',
+      },
+      Sir: {
+        id: 'option.personTitles.Sir',
+        defaultMessage: 'Sir',
+      },
+    }),
+  },
+};